refactor measure profile, move profile webservice into angular service
