+++ geoportailv3/templates/index.html
@@ -280,7 +280,7 @@
          appModule.constant('logoutUrl', "${request.route_url('logout')}");
          appModule.constant('getuserinfoUrl', "${request.route_url('getuserinfo')}");
          appModule.constant('elevationServiceUrl', "${request.route_url('raster')}");
-         appModule.constant('profilejsonUrl', "${request.route_url('profile.json')}");
+         appModule.constant('profileServiceUrl', "${request.route_url('profile.json')}");
          appModule.constant('searchServiceUrl', "${request.route_url('fulltextsearch')}");
          appModule.constant('printServiceUrl', "${request.route_url('printproxy')}");
          appModule.constant('shorturlServiceUrl', "${request.route_url('shortener_create')}");
